Presentation Task: - remove redundant class of recyclerView - move custom views in customDailyCalendarViewGroup package
