+++ app/src/main/res/layout/fragment_to_do_list.xml
@@ -69,14 +69,14 @@
                 android:contentDescription="@string/container_for_daily_calendar"
                 >
 
-                <com.example.todolist.presantation.toDoList.CustomDailyCalendarViewGroup
+                <com.example.todolist.presantation.toDoList.customDailyCalendarViewGroup.CustomDailyCalendarViewGroup
                     android:id="@+id/daily_calendar_view_group"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
                     app:eachRowHeight="20dp"
                     app:minimumPercentage="0.25"
                     app:numberOfRows="24">
-                </com.example.todolist.presantation.toDoList.CustomDailyCalendarViewGroup>
+                </com.example.todolist.presantation.toDoList.customDailyCalendarViewGroup.CustomDailyCalendarViewGroup>
             </HorizontalScrollView>
         </ScrollView>
